In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -12,26 +12,26 @@
 
 int main(void)
 {
-	long long pro1, pro2 = 1, pro3 = 0;
+	int pro1, pro2 = 1, pro3 = 0;
 
 	while (pro3 <= 20365011074)
 	{
 		if (pro3 == 20365011074)
 		{
-			printf("%lld\n", pro3);
+			printf("%u\n", pro3);
 			break;
 		}
 		pro1 = pro2 + pro3;
-		printf("%lld, ", pro1);
+		printf("%u, ", pro1);
 		pro3 = pro1 + pro2;
 		if (pro3 != 20365011074)
 		{
-			printf("%lld, ", pro3);
+			printf("%u, ", pro3);
 		}
 		if (pro3 != 20365011074)
 		{
 		pro2 = pro3 + pro1;
-		printf("%lld, ", pro2);
+		printf("%u, ", pro2);
 		}
 	}
 	return (0);
